marllib/patch/hns/mujoco-worldgen/requirements.txt
click==7.0
xmltodict==0.12.0
scipy==1.3.1
gym==0.10.8
jsonnet==0.11.2
numpy-stl==2.10.1
mujoco-py<1.50.2,>=1.50.1
